backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
uvicorn==0.24.0
pyjwt==2.8.0
psycopg2-binary>=2.9.0
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
slowapi==0.1.8
passlib==1.7.4
bcrypt==4.0.1